fix: checklist flow
